project 14/2

+++ controllers/cards.js
@@ -1,76 +1,57 @@
 const Card = require('../models/card');
+const ForbiddenError = require('../errors/forbidden-error');
+const NotFoundError = require('../errors/not-found-err');
 
-module.exports.getCards = (req, res) => {
+module.exports.getCards = (req, res, next) => {
   Card.find({})
     .then((cards) => res.send(cards))
-    .catch(() => res.status(500).send({ message: 'На сервере произошла ошибка' }));
+    .catch(next);
 };
-module.exports.createCard = (req, res) => {
+module.exports.createCard = (req, res, next) => {
   const { name, link } = req.body;
   const owner = req.user._id;
   Card.create({ name, link, owner })
     .then((card) => res.status(201).send(card))
-    .catch((err) => {
-      if (err.name === 'ValidationError') {
-        return res.status(400).send({ message: 'Переданы некорректные данные при создании карточки' });
-      }
-      return res.status(500).send({ message: 'Ошибка на стороне сервера' });
-    });
+    .catch(next);
 };
-module.exports.deleteCardById = (req, res) => {
-  Card.findByIdAndDelete(req.params.cardId)
-    .orFail(new Error('NotValidId'))
-    .then(() => {
-      res.send({ message: 'Вы удалили карточку' });
-    })
-    .catch((err) => {
-      if (err.message === 'NotValidId') {
-        res.status(404).send({ message: 'Карточка с указанным _id не найдена' });
-      } else if (err.name === 'CastError') {
-        res.status(400).send({ message: 'Переданы некорректные данные для удаления карточки' });
-      } else {
-        res.status(500).send({ message: 'Ошибка на стороне сервера' });
+module.exports.deleteCardById = (req, res, next) => {
+  Card.findById(req.params.cardId)
+    .orFail(new NotFoundError('Карточка с указанным _id не найдена'))
+    .then((card) => {
+      if (`${card.owner}` !== req.user._id) {
+        throw new ForbiddenError('Вы не можете удалить карточку другого пользователя');
       }
-    });
+      Card.findByIdAndDelete(req.params.cardId)
+        .orFail(new NotFoundError('Карточка с указанным _id не найдена'))
+        .then(() => {
+          res.send({ message: 'успешно' });
+        })
+        .catch(next);
+    })
+    .catch(next);
 };
 
-module.exports.likeCard = (req, res) => {
+module.exports.likeCard = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $addToSet: { likes: req.user._id } },
     { new: true },
   )
-    .orFail(new Error('NotValidId'))
+    .orFail(new NotFoundError('Карточка с указанным _id не найдена'))
     .then((card) => {
       res.send(card);
     })
-    .catch((err) => {
-      if (err.message === 'NotValidId') {
-        res.status(404).send({ message: 'Карточка с указанным _id не найдена' });
-      } else if (err.name === 'CastError') {
-        res.status(400).send({ message: 'Переданы некорректные данные для постановки лайка' });
-      } else {
-        res.status(500).send({ message: 'Ошибка на стороне сервера' });
-      }
-    });
+    .catch(next);
 };
-module.exports.dislikeCard = (req, res) => {
+module.exports.dislikeCard = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $pull: { likes: req.user._id } },
     { new: true },
   )
-    .orFail(new Error('NotValidId'))
+    .orFail(new NotFoundError('Карточка с указанным _id не найдена'))
     .then((card) => {
       res.send(card);
     })
-    .catch((err) => {
-      if (err.message === 'NotValidId') {
-        res.status(404).send({ message: 'Карточка с указанным _id не найдена' });
-      } else if (err.name === 'CastError') {
-        res.status(400).send({ message: 'Переданы некорректные данные для снятии лайка' });
-      } else {
-        res.status(500).send({ message: 'Ошибка на стороне сервера' });
-      }
-    });
+    .catch(next);
 };
